Refactoring: move implementation of potentially static factories to the cpp file
